Rename bisq.httpapi package to bisq.api.http
