+++ http-api/src/testIntegration/java/bisq/api/http/SwaggerIT.java
@@ -1,4 +1,4 @@
-package bisq.httpapi;
+package bisq.api.http;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
