+++ http-api/src/testIntegration/resources/META-INF/services/org.jboss.arquillian.core.spi.LoadableExtension
@@ -1,1 +1,1 @@
-bisq.httpapi.arquillian.CubeLoggerExtension
+bisq.api.http.arquillian.CubeLoggerExtension
